refactor(inference): break down files
